requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
# For production Postgres, uncomment:
# psycopg2-binary==2.9.9
# alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.5.2

# HTTP client (for Instagram + AI API calls)
httpx==0.27.2

# Dev/Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # also used as test client
